#5463 feat(refactor): Renamed ResultList component, removed extraneous files and dependencies
