+++ package.json
@@ -44,9 +44,7 @@
   "dependencies": {
     "classnames": "^2.2.6",
     "lodash.throttle": "^4.1.1",
-    "react-cookie": "^4.0.3",
-    "react-paginate": "^6.3.2",
-    "react-solr-connector": "^1.1.1"
+    "react-cookie": "^4.0.3"
   },
   "devDependencies": {
     "@babel/cli": "^7.7.7",
