+++ src/assets/styles/30-components/__manifest.scss
@@ -15,10 +15,9 @@
 @import './src/components/Nav/nav-item';
 @import './src/components/Nav/nav-list';
 @import './src/components/PageHeader/page-header';
-@import './src/components/ResultItem/result-item';
+@import './src/components/ResultsItem/results-item';
+@import './src/components/ResultsItem/results-item';
 @import './src/components/SearchForm/search-form';
 @import './src/components/SearchPane/search-pane';
-@import './src/components/WellcomeCollectionBanner/wellcome-collection-banner';
-@import './src/Components/Logo/logo';
-@import './src/components/ResultItem/result-item';
 @import './src/components/Section/section';
+@import './src/components/WellcomeCollectionBanner/wellcome-collection-banner';
