+++ src/components/ResultsItem/_results-item.scss
