+++ src/components/index.ts
@@ -8,18 +8,16 @@ export { Grid, GridCell } from 'Grid/Grid';
 export { Hamburger } from 'Hamburger/Hamburger';
 export { Header } from 'Header/Header';
 export { Icon } from 'Icon/Icon';
+export { Logo } from 'Logo/Logo';
 export { NavItem } from 'Nav/NavItem';
 export { NavLinks } from 'Nav/NavLinks';
-export { RouterLinkWrapper } from 'RouterLinkWrapper/RouterLinkWrapper';
-export { WellcomeCollectionBanner } from 'WellcomeCollectionBanner/WellcomeCollectionBanner';
-export { Logo } from 'Logo/Logo';
 export { PageHeader } from 'PageHeader/PageHeader';
-export { ResultItem } from 'ResultItem/ResultItem';
-export { ResultsList } from 'ResultsList/ResultsList';
+export { ResultsItem } from 'ResultsItem/ResultsItem';
+export { RouterLinkWrapper } from 'RouterLinkWrapper/RouterLinkWrapper';
 export { SearchForm } from 'SearchForm/SearchForm';
 export { SearchPane } from 'SearchPane/SearchPane';
-export { SearchQueryFetch } from 'SearchQueryFetch/SearchQueryFetch';
 export { Section } from 'Section/Section';
+export { WellcomeCollectionBanner } from 'WellcomeCollectionBanner/WellcomeCollectionBanner';
 
 // context components
 export { NavContext, NavContextProvider } from 'NavContext/NavContext';
